Refactor meeting config and CLI into package modules
