+++ backend/ai_meeting/__main__.py
@@ -1,14 +1,7 @@
-"""`python -m backend.ai_meeting` 用の暫定ラッパー。"""
+"""`python -m backend.ai_meeting` 用のエントリーポイント。"""
 from __future__ import annotations
 
-from . import get_main
-
-
-def main() -> None:
-    """旧 `main()` を呼び出す。"""
-
-    legacy_main = get_main()
-    legacy_main()
+from .cli import main
 
 
 if __name__ == "__main__":
